# mathext: speed up Digamma for large negative arguments