--- a/CONTRIBUTORS
+++ b/CONTRIBUTORS
@@ -41,6 +41,7 @@ Jeremy Atkinson <jchatkinson@gmail.com>
 Jonathan J Lawlor <jonathan.lawlor@gmail.com>
 Jonathan Schroeder <jd.schroeder@gmail.com>
 Joseph Watson <jtwatson@linux-consulting.us>
+Josh Wilson <josh.craig.wilson@gmail.com>
 Julien Roland <juroland@gmail.com>
 Kent English <kent.english@gmail.com>
 Konstantin Shaposhnikov <k.shaposhnikov@gmail.com>