Before the Digamma function could be quite slow for large negative
arguments because it had to do use the recurrence relation thousands
of times. This avoids that by instead using the reflection formula. It
also adds some explicit checks for special cases (Inf, NaN, poles).

For example, running the following code:

```go
package main

import (
	"fmt"
	"time"
	"gonum.org/v1/gonum/mathext"
)

func main () {
	start := time.Now()
	for i := 0; i < 10; i++ {
		mathext.Digamma(-1.1111111111e9)
	}
	fmt.Printf("%s elapsed\n", time.Since(start))
}
```
gives me `40.625090145s elapsed` on master and `8.615µs elapsed` on this branch.