@person142 2017-11-16
@kortschak would you prefer the fixes be squashed into one commit or kept separate?