@kortschak 2017-11-16
There's nothing egregious here, so I reckon a squash of fixes works. Alternatively leave them separate and make a separate PR for the addition to A+C.