@btracey 2017-11-17
Algorithm changes LGTM.