@person142 2017-11-17
Review comments (hopefully) addressed.

Benchmarks: on this branch I get

```
go test -bench=BenchmarkDigamma
goos: darwin
goarch: amd64
pkg: gonum.org/v1/gonum/mathext
BenchmarkDigamma-8   	300000000	         4.82 ns/op
PASS
ok  	gonum.org/v1/gonum/mathext	6.297s
```

On master it times out:
```
go test -bench=BenchmarkDigamma
SIGQUIT: quit
PC=0x1105204 m=0 sigcode=0

[snip]

*** Test killed with quit: ran too long (10m0s).
FAIL	gonum.org/v1/gonum/mathext	600.011s
```