@kortschak 2017-11-17
That's a pretty compelling benchmark! I think what has happened though is that it's timed out on the tests and failed to get to the bench. You can skip all the tests by adding `-run ^$` to regex out the full suite.

I get the following:
```
name       old time/op  new time/op  delta
Digamma-8   8.37s ± 0%   0.00s ± 2%  -100.00%  (p=0.000 n=9+10)
```

(The 0s is ~8.3ns).